# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DuckyAPI",
    "description": "A customer facing api for WildDuck",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "Package": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "object",
            "example": "5d49e11f600a423ffc0b1297",
            "description": "Unique id for this package",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "example": "Small",
            "description": "Display name to use for this package"
          },
          "quota": {
            "type": "number",
            "example": 1073741824,
            "description": "Storage quota in bytes, 0 is unlimited"
          },
          "maxSend": {
            "type": "number",
            "example": 200,
            "description": "Max send quota for accounts created by this user, 0 is unlimited"
          },
          "maxReceive": {
            "type": "number",
            "example": 1000,
            "description": "Max recieve quota for accounts created by this user, 0 is unlimited"
          },
          "maxForward": {
            "type": "number",
            "example": 100,
            "description": "Max forward quota for accounts created by this user, 0 is unlimited"
          }
        },
        "required": ["name"]
      },
...
    }
  },
  "paths": {
...
    "/packages": {
      "get": {
        "operationId": "getPackages",
        "summary": "[Admin only] Get a list of packages",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of packages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Package" }
                }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" }
        },
        "tags": ["Packages"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request GET \\\n  --url http://localhost:3000/packages \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"GET\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/packages\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"GET\", \"http://localhost:3000/packages\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"GET\", \"/packages\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/packages\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"GET\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.get(\"http://localhost:3000/packages\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/packages\");\nvar request = new RestRequest(Method.GET);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"GET\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/packages\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      },
      "post": {
        "operationId": "createPackage",
        "summary": "[Admin only] Create package",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Package" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created package",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Package" }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" }
        },
        "tags": ["Packages"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request POST \\\n  --url http://localhost:3000/packages \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN' \\\n  --header 'content-type: application/json' \\\n  --data '{\"name\":\"Small\",\"quota\":1073741824,\"maxSend\":200,\"maxReceive\":1000,\"maxForward\":100}'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"POST\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/packages\",\n  \"headers\": {\n    \"content-type\": \"application/json\",\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.write(JSON.stringify({\n  name: 'Small',\n  quota: 1073741824,\n  maxSend: 200,\n  maxReceive: 1000,\n  maxForward: 100\n}));\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = JSON.stringify({\n  \"name\": \"Small\",\n  \"quota\": 1073741824,\n  \"maxSend\": 200,\n  \"maxReceive\": 1000,\n  \"maxForward\": 100\n});\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"POST\", \"http://localhost:3000/packages\");\nxhr.setRequestHeader(\"content-type\", \"application/json\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\npayload = \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\"\n\nheaders = {\n    'content-type': \"application/json\",\n    'authorization': \"Bearer REPLACE_BEARER_TOKEN\"\n    }\n\nconn.request(\"POST\", \"/packages\", payload, headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/packages\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"POST\",\n  CURLOPT_POSTFIELDS => \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\",\n    \"content-type: application/json\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.post(\"http://localhost:3000/packages\")\n  .header(\"content-type\", \"application/json\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .body(\"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/packages\");\nvar request = new RestRequest(Method.POST);\nrequest.AddHeader(\"content-type\", \"application/json\");\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nrequest.AddParameter(\"application/json\", \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\", ParameterType.RequestBody);\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/packages\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"content-type: application/json\");\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\ncurl_easy_setopt(hnd, CURLOPT_POSTFIELDS, \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\");\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
    "/packages/{id}": {
      "put": {
        "operationId": "updatePackage",
        "summary": "[Admin only] Update package",
        "description": "Will also update quota for existing users, except if you modified the users quota manually.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "example": "5d49e11f600a423ffc0b1297",
            "description": "Unique id for the package",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Package" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated package",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Package" }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" }
        },
        "tags": ["Packages"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request PUT \\\n  --url http://localhost:3000/packages/%7Bid%7D \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN' \\\n  --header 'content-type: application/json' \\\n  --data '{\"name\":\"Small\",\"quota\":1073741824,\"maxSend\":200,\"maxReceive\":1000,\"maxForward\":100}'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"PUT\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/packages/%7Bid%7D\",\n  \"headers\": {\n    \"content-type\": \"application/json\",\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.write(JSON.stringify({\n  name: 'Small',\n  quota: 1073741824,\n  maxSend: 200,\n  maxReceive: 1000,\n  maxForward: 100\n}));\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = JSON.stringify({\n  \"name\": \"Small\",\n  \"quota\": 1073741824,\n  \"maxSend\": 200,\n  \"maxReceive\": 1000,\n  \"maxForward\": 100\n});\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"PUT\", \"http://localhost:3000/packages/%7Bid%7D\");\nxhr.setRequestHeader(\"content-type\", \"application/json\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\npayload = \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\"\n\nheaders = {\n    'content-type': \"application/json\",\n    'authorization': \"Bearer REPLACE_BEARER_TOKEN\"\n    }\n\nconn.request(\"PUT\", \"/packages/%7Bid%7D\", payload, headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/packages/%7Bid%7D\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"PUT\",\n  CURLOPT_POSTFIELDS => \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\",\n    \"content-type: application/json\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.put(\"http://localhost:3000/packages/%7Bid%7D\")\n  .header(\"content-type\", \"application/json\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .body(\"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/packages/%7Bid%7D\");\nvar request = new RestRequest(Method.PUT);\nrequest.AddHeader(\"content-type\", \"application/json\");\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nrequest.AddParameter(\"application/json\", \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\", ParameterType.RequestBody);\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/packages/%7Bid%7D\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"content-type: application/json\");\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\ncurl_easy_setopt(hnd, CURLOPT_POSTFIELDS, \"{\\\"name\\\":\\\"Small\\\",\\\"quota\\\":1073741824,\\\"maxSend\\\":200,\\\"maxReceive\\\":1000,\\\"maxForward\\\":100}\");\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      },
      "delete": {
        "operationId": "deletePackage",
        "summary": "[Admin only] Delete package",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "example": "5d49e11f600a423ffc0b1297",
            "description": "Unique id for the package",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted package",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Package" }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" }
        },
        "tags": ["Packages"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request DELETE \\\n  --url http://localhost:3000/packages/%7Bid%7D \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"DELETE\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/packages/%7Bid%7D\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"DELETE\", \"http://localhost:3000/packages/%7Bid%7D\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"DELETE\", \"/packages/%7Bid%7D\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/packages/%7Bid%7D\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"DELETE\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.delete(\"http://localhost:3000/packages/%7Bid%7D\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/packages/%7Bid%7D\");\nvar request = new RestRequest(Method.DELETE);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/packages/%7Bid%7D\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
...
  }
}